requirements.txt
aiofiles==24.1.0
aiogram==3.22.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
aiosqlite==0.21.0
annotated-types==0.7.0
attrs==25.3.0
certifi==2025.8.3
frozenlist==1.7.0
greenlet==3.0.3
idna==3.10
magic-filter==1.0.12
multidict==6.6.4
numpy==2.3.3
pandas==2.3.2
propcache==0.3.2
pydantic==2.11.9
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
six==1.17.0
SQLAlchemy==2.0.43
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
yarl==1.20.1
